+++ src/storage/storage_transformer/performance_metrics.rs
@@ -103,7 +103,7 @@ struct PerformanceMetricsStorageTransformerImpl<'a, TStorage: ?Sized> {
     transformer: &'a PerformanceMetricsStorageTransformer,
 }
 
-impl<TStorage: ReadableStorageTraits + ?Sized> ReadableStorageTraits
+impl<TStorage: ?Sized + ReadableStorageTraits> ReadableStorageTraits
     for PerformanceMetricsStorageTransformerImpl<'_, TStorage>
 {
     fn get(&self, key: &StoreKey) -> Result<Vec<u8>, StorageError> {
@@ -149,7 +149,7 @@ impl<TStorage: ReadableStorageTraits + ?Sized> ReadableStorageTraits
     }
 }
 
-impl<TStorage: ListableStorageTraits + ?Sized> ListableStorageTraits
+impl<TStorage: ?Sized + ListableStorageTraits> ListableStorageTraits
     for PerformanceMetricsStorageTransformerImpl<'_, TStorage>
 {
     fn list(&self) -> Result<StoreKeys, StorageError> {
@@ -165,7 +165,7 @@ impl<TStorage: ListableStorageTraits + ?Sized> ListableStorageTraits
     }
 }
 
-impl<TStorage: WritableStorageTraits + ?Sized> WritableStorageTraits
+impl<TStorage: ?Sized + WritableStorageTraits> WritableStorageTraits
     for PerformanceMetricsStorageTransformerImpl<'_, TStorage>
 {
     fn set(&self, key: &StoreKey, value: &[u8]) -> Result<(), StorageError> {
@@ -206,7 +206,7 @@ impl<TStorage: WritableStorageTraits + ?Sized> WritableStorageTraits
     }
 }
 
-impl<TStorage: ReadableStorageTraits + WritableStorageTraits + ?Sized> ReadableWritableStorageTraits
+impl<TStorage: ?Sized + ReadableStorageTraits + WritableStorageTraits> ReadableWritableStorageTraits
     for PerformanceMetricsStorageTransformerImpl<'_, TStorage>
 {
 }